re-introduce updateAudioSourceFunction
re-introduce function that is part of the AOAv2 audio feature that was removed first in Android pie.

code seen here in AOSP repo
https://cs.android.com/android/_/android/platform/frameworks/base/+/412d5e191b7ad98a4dbc818c2e46954c94784872:services/usb/java/com/android/server/usb/UsbDeviceManager.java;drc=371a3b879ba82bbe5a4d914328a20659131d0220;bpv=1;bpt=0;dlc=0be310a3039870d4e5674d356f5ca2ffb06c1309

This was a challenge to find the related change, because this does not appear in any commit. It only exists in the comparison between two neighboring commits. A suspicious finding...